==> Notebooks/1_Text_preprocessing.ipynb <==
# Mount Google Drive
from google.colab import drive
drive.mount('/content/drive')

# Download NLTK and spaCy models
import nltk
nltk.download('punkt')
nltk.download('stopwords')

import spacy
spacy.cli.download("en_core_web_sm")

# Import libraries
!pip install nltk spacy datasketch unidecode simhash

import pandas as pd
import re
import numpy as np
from unidecode import unidecode
from simhash import Simhash
from datasketch import MinHash, MinHashLSH

# Load spaCy
import spacy
nlp = spacy.load("en_core_web_sm")


# Load Data from Gdrive
file_path = "/content/drive/MyDrive/NLP - Group Project/recipe_dataset.csv"

# Load the dataset
df = pd.read_csv(file_path)
print("Original shape:", df.shape)

# Sample data
df_sample = df.sample(1000000, random_state=42).copy()
print("Sample shape:", df_sample.shape)

# Fill missing value
df_sample['ingredients'] = df_sample['ingredients'].fillna("").astype(str)
df_sample['directions'] = df_sample['directions'].fillna("").astype(str)

# Text cleaning
def simple_clean(text):
    text = text.lower()
    text = unidecode(text)
    text = re.sub(r'[^a-z\s]', '', text)
    return re.sub(r'\s+', ' ', text).strip()

df_sample['clean_ingredients'] = df_sample['ingredients'].apply(simple_clean)
df_sample['clean_directions'] = df_sample['directions'].apply(simple_clean)
df_sample['clean_text'] = df_sample['clean_ingredients'] + ' ' + df_sample['clean_directions']

SimHash Deduplication

def simhash_text(text, max_words=100):
    tokens = text.split()[:max_words]
    return Simhash(' '.join(tokens)).value

df_sample['simhash'] = df_sample['clean_text'].apply(simhash_text)
df_sample = df_sample.drop_duplicates(subset='simhash').reset_index(drop=True)
print("After SimHash Deduplication Deduplication:", df_sample.shape)


MinHash Deduplication

def minhash_signature(text, num_perm=128):
    m = MinHash(num_perm=num_perm)
    for word in text.split()[:100]:
        m.update(word.encode('utf8'))
    return m

lsh = MinHashLSH(threshold=0.9, num_perm=128)
minhashes = {}

for i, row in df_sample.iterrows():
    m = minhash_signature(row['clean_text'])
    minhashes[i] = m
    lsh.insert(i, m)

to_drop = set()
for i in minhashes:
    results = lsh.query(minhashes[i])
    for r in results:
        if i != r:
            to_drop.add(r)

df_sample = df_sample.drop(index=to_drop).reset_index(drop=True)
print("After MinHash Deduplication:", df_sample.shape)

Save to CSV

# Save deduplicated data after SimHash and MinHash
output_path = "/content/drive/MyDrive/NLP - Group Project/preprocessed_recipes.csv"
df_sample.to_csv(output_path, index=False)
print("✅ Saved cleaned dataset to:", output_path)

df_sample.head(10)
